+++ changelog.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.9.0] 2021-11-21
 
 - Change Parser.Group signature, removing items argument
 - Replace Parser.Optional and Parser.Required with Parser.NamedArg